convert all width and height into double data type which more make sense. added top and left for WorkItem to place position in workspace. method of save xaml into output file cannot be use anymore because it does not keep binding. So direct databinding to WorkItem (together with drawing element) is the correct direction to go.

+++ HLGranite.Drawing/BaseItem.cs
@@ -7,8 +7,8 @@ namespace HLGranite.Drawing
 {
     public partial class BaseItem
     {
-        protected int widthField;
-        protected int heightField;
+        protected double widthField;
+        protected double heightField;
 
         public BaseItem()
             : base()
